backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
psutil==5.9.5
supabase==2.0.0
flask-cors==4.0.0
